Fix require cycle
