+++ src/libs/Notification/PushNotification/index.native.js
@@ -2,34 +2,9 @@ import _ from 'underscore';
 import {AppState} from 'react-native';
 import {UrbanAirship, EventType, iOS} from 'urbanairship-react-native';
 import lodashGet from 'lodash/get';
-import Onyx from 'react-native-onyx';
-import DeviceInfo from 'react-native-device-info';
-import ONYXKEYS from '../../../ONYXKEYS';
 import Log from '../../Log';
 import NotificationType from './NotificationType';
 import * as User from '../../actions/User';
-
-const deviceID = DeviceInfo.getDeviceId();
-
-let isUserOptedInToPushNotifications = false;
-Onyx.connect({
-    key: ONYXKEYS.NVP_PUSH_NOTIFICATIONS_ENABLED,
-    callback: (val) => {
-        const pushNotificationOptInRecord = lodashGet(val, deviceID, []);
-        const mostRecentNVPValue = _.last(pushNotificationOptInRecord);
-        if (!_.has(mostRecentNVPValue, 'isEnabled')) {
-            return;
-        }
-        isUserOptedInToPushNotifications = mostRecentNVPValue.isEnabled;
-    },
-});
-
-/**
- * @returns {Boolean}
- */
-function isUserOptedIn() {
-    return isUserOptedInToPushNotifications;
-}
 
 const notificationEventActionMap = {};
 
@@ -78,7 +53,7 @@ function refreshNotificationOptInStatus() {
     UrbanAirship.getNotificationStatus()
         .then((notificationStatus) => {
             const isOptedIn = notificationStatus.airshipOptIn && notificationStatus.systemEnabled;
-            if (isOptedIn === isUserOptedInToPushNotifications) {
+            if (isOptedIn === User.isUserOptedIntoPushNotifications()) {
                 return;
             }
 
@@ -98,7 +73,7 @@ function init() {
     // Setup event listeners
     UrbanAirship.addListener(EventType.PushReceived, (notification) => {
         // By default, refresh notification opt-in status to true if we receive a notification
-        if (!isUserOptedInToPushNotifications) {
+        if (!User.isUserOptedIntoPushNotifications()) {
             User.setPushNotificationOptInStatus(true);
         }
 
@@ -220,7 +195,6 @@ function clearNotifications() {
 }
 
 export default {
-    isUserOptedIn,
     init,
     register,
     deregister,
